sdk: silo cleanup - extra helper methods to utilities file
