+++ projects/sdk/src/lib/silo.ts
@@ -1,24 +1,18 @@
-import { ethers, BigNumber, ContractTransaction } from "ethers";
-import { ERC20Token, Token } from "src/classes/Token";
+import { BigNumber, ContractTransaction } from "ethers";
+import { Token } from "src/classes/Token";
 import { BeanstalkSDK, DataSource } from "./BeanstalkSDK";
 import EventProcessor from "src/lib/events/processor";
-import { EIP712Domain, EIP712TypedData, Permit } from "./permit";
-import {
-  CrateSortFn,
-  DepositTokenPermitMessage,
-  DepositTokensPermitMessage,
-  sortCratesBySeason,
-  parseWithdrawalCrates
-} from "./silo/utils";
+import { EIP712TypedData } from "./permit";
+import * as utils from "./silo/utils";
+import * as permitUtils from "./silo/utils.permit";
 import { TokenValue } from "src/classes/TokenValue";
 import { MAX_UINT256 } from "src/constants";
-import { assert } from "src/utils";
 import { DepositBuilder } from "./silo/DepositBuilder";
 import { DepositOperation } from "./silo/DepositOperation";
 import { Withdraw } from "./silo/Withdraw";
 import { Claim } from "./silo/Claim";
 import { FarmToMode } from "./farm";
-import { DepositCrate, WithdrawalCrate, Crate, TokenSiloBalance } from "./silo/types";
+import { DepositCrate, TokenSiloBalance, DepositTokenPermitMessage, DepositTokensPermitMessage } from "./silo/types";
 
 export class Silo {
   static sdk: BeanstalkSDK;
@@ -35,6 +29,22 @@ export class Silo {
     this.depositBuilder = new DepositBuilder(sdk);
     this.siloWithdraw = new Withdraw(sdk);
     this.siloClaim = new Claim(sdk);
+  }
+
+  /**
+   * Mowing adds Grown Stalk to stalk balance
+   * @param _account
+   */
+  async mow(_account?: string): Promise<ContractTransaction> {
+    const account = _account ? _account : await Silo.sdk.getAccount();
+    return Silo.sdk.contracts.beanstalk.update(account);
+  }
+
+  /**
+   * Claims Earned Beans, Earned Stalk, Plantable Seeds and also mows any Grown Stalk
+   */
+  async plant(): Promise<ContractTransaction> {
+    return Silo.sdk.contracts.beanstalk.plant();
   }
 
   /**
@@ -126,190 +136,6 @@ export class Silo {
   }
 
   /**
-   * Sort the incoming map so that tokens are ordered in the same order
-   * they appear on the Silo Whitelist.
-   *
-   * @note the Silo Whitelist is sorted by the order in which tokens were
-   * whitelisted in Beanstalk. Unclear if the ordering shown on the
-   * Beanstalk UI will change at some point in the future.
-   */
-  private _sortTokenMapByWhitelist<T extends any>(map: Map<Token, T>) {
-    const whitelist = Silo.sdk.tokens.siloWhitelist;
-    const copy = new Map<Token, T>(map);
-    const ordered = new Map<Token, T>();
-    // by default, order by whitelist
-    whitelist.forEach((token) => {
-      const v = copy.get(token);
-      if (v) {
-        ordered.set(token, v);
-        copy.delete(token);
-      }
-    });
-    // add remaining tokens
-    copy.forEach((_, token) => {
-      ordered.set(token, copy.get(token)!);
-    });
-    return ordered;
-  }
-
-  //////////////////////// WHITELIST ////////////////////////
-
-  /**
-   * Return a list of tokens that are currently whitelisted in the Silo.
-   *
-   * @todo Check if the subgraph removes `WhitelistToken` entities if a
-   *       token is de-whitelisted.
-   * @todo Get name, decimals since these are ERC20 tokens.
-   */
-  public async getWhitelist(options?: { source: DataSource.LEDGER } | { source: DataSource.SUBGRAPH }) {
-    const source = Silo.sdk.deriveConfig("source", options);
-    if (source === DataSource.SUBGRAPH) {
-      const query = await Silo.sdk.queries.getSiloWhitelist();
-      return query.whitelistTokens.map((e) => ({
-        token: e.token,
-        stalk: parseInt(e.stalk),
-        seeds: parseInt(e.seeds) / 1e4
-      }));
-    }
-    throw new Error(`Unsupported source: ${source}`);
-  }
-
-  //////////////////////// BALANCES ////////////////////////
-
-  private _parseWithdrawalCrates = parseWithdrawalCrates;
-
-  private _makeTokenSiloBalance(): TokenSiloBalance {
-    return {
-      deposited: {
-        amount: TokenValue.ZERO,
-        bdv: TokenValue.ZERO,
-        crates: [] as DepositCrate[]
-      },
-      withdrawn: {
-        amount: TokenValue.ZERO,
-        crates: [] as WithdrawalCrate[]
-      },
-      claimable: {
-        amount: TokenValue.ZERO,
-        crates: [] as WithdrawalCrate[]
-      }
-    };
-  }
-
-  /**
-   * Calculate the amount Stalk grown since `depositSeason`.
-   * Depends on the `currentSeason` and the `depositSeeds` awarded
-   * for a particular deposit.
-   *
-   * @param currentSeason
-   * @param depositSeason
-   * @param depositSeeds
-   * @returns TokenValue<STALK>
-   */
-  public calculateGrownStalk(currentSeason: ethers.BigNumberish, depositSeason: ethers.BigNumberish, depositSeeds: TokenValue): TokenValue {
-    const deltaSeasons = ethers.BigNumber.from(currentSeason).sub(depositSeason);
-    assert(deltaSeasons.gte(0), "Silo: Cannot calculate grown stalk when `currentSeason < depositSeason`.");
-    return Silo.STALK_PER_SEED_PER_SEASON.mul(depositSeeds).mul(deltaSeasons.toNumber());
-  }
-
-  /**
-   * Create a new Deposit Crate object.
-   *
-   * @param token Token contained within the crate
-   * @param _season The season of deposit
-   * @param _amount The amount of deposit
-   * @param _bdv The bdv of deposit
-   * @param currentSeason The current season, for calculation of grownStalk.
-   * @returns DepositCrate<TokenValue>
-   */
-  public makeDepositCrate(
-    token: Token,
-    _season: string | number,
-    _amount: string,
-    _bdv: string,
-    currentSeason: ethers.BigNumberish
-  ): DepositCrate<TokenValue> {
-    // Crate
-    const season = ethers.BigNumber.from(_season);
-    const amount = token.fromBlockchain(_amount);
-
-    // Deposit-specific
-    const bdv = Silo.sdk.tokens.BEAN.fromBlockchain(_bdv);
-    const seeds = token.getSeeds(bdv);
-    const baseStalk = token.getStalk(bdv);
-    const grownStalk = this.calculateGrownStalk(currentSeason, season, seeds);
-    const stalk = baseStalk.add(grownStalk);
-
-    return {
-      season,
-      amount,
-      bdv,
-      stalk,
-      baseStalk,
-      grownStalk,
-      seeds
-    };
-  }
-
-  /**
-   * Apply a Deposit to a TokenSiloBalance.
-   * @note expects inputs to be stringified (no decimals).
-   */
-  private _applyDeposit(
-    state: TokenSiloBalance["deposited"],
-    token: Token,
-    rawCrate: {
-      season: string | number;
-      amount: string;
-      bdv: string;
-    },
-    currentSeason: ethers.BigNumberish
-  ) {
-    const crate = this.makeDepositCrate(token, rawCrate.season, rawCrate.amount, rawCrate.bdv, currentSeason);
-
-    state.amount = state.amount.add(crate.amount);
-    state.bdv = state.bdv.add(crate.bdv);
-    state.crates.push(crate);
-
-    return crate;
-  }
-
-  /**
-   * Apply a Deposit to a TokenSiloBalance.
-   *
-   * @note expects inputs to be stringified (no decimals).
-   */
-  private _applyWithdrawal(
-    state: TokenSiloBalance["withdrawn" | "claimable"],
-    token: Token,
-    rawCrate: {
-      season: string | number;
-      amount: string;
-    }
-  ) {
-    const season = BigNumber.from(rawCrate.season);
-    const amount = token.amount(rawCrate.amount);
-
-    const crate: Crate<TokenValue> = {
-      season: season,
-      amount: amount
-    };
-
-    state.amount = state.amount.add(amount);
-    state.crates.push(crate);
-
-    return crate;
-  }
-
-  private _sortCrates(state: TokenSiloBalance["deposited" | "withdrawn" | "claimable"]) {
-    state.crates = state.crates.sort(
-      (a, b) => a.season.sub(b.season).toNumber() // sort by season asc
-    );
-  }
-
-  //////////////////////// Balances & Amounts ////////////////////////
-
-  /**
    * Return the Farmer's balance of a single whitelisted token.
    */
   public async getBalance(
@@ -325,7 +151,7 @@ export class Silo {
 
     ///  SETUP
     const whitelist = Silo.sdk.tokens.siloWhitelist;
-    const balance: TokenSiloBalance = this._makeTokenSiloBalance();
+    const balance: TokenSiloBalance = utils.makeTokenSiloBalance();
 
     if (source === DataSource.LEDGER) {
       // Fetch and process events.
@@ -350,23 +176,23 @@ export class Silo {
           };
           // Update the total deposited of this token
           // and return a parsed crate object
-          this._applyDeposit(balance.deposited, _token, rawCrate, currentSeason);
+          utils.applyDeposit(balance.deposited, _token, rawCrate, currentSeason);
         }
 
-        this._sortCrates(balance.deposited);
+        utils.sortCrates(balance.deposited);
       }
 
       // Handle withdrawals
       {
         const _crates = withdrawals.get(_token);
         if (_crates) {
-          const { withdrawn, claimable } = this._parseWithdrawalCrates(_token, _crates, seasonBN);
+          const { withdrawn, claimable } = utils.parseWithdrawalCrates(_token, _crates, seasonBN);
 
           balance.withdrawn = withdrawn;
           balance.claimable = claimable;
 
-          this._sortCrates(balance.withdrawn);
-          this._sortCrates(balance.claimable);
+          utils.sortCrates(balance.withdrawn);
+          utils.sortCrates(balance.claimable);
         }
       }
 
@@ -383,9 +209,9 @@ export class Silo {
       if (!query.farmer) return balance;
 
       const { deposited, withdrawn, claimable } = query.farmer!;
-      deposited.forEach((crate) => this._applyDeposit(balance.deposited, _token, crate, currentSeason));
-      withdrawn.forEach((crate) => this._applyWithdrawal(balance.withdrawn, _token, crate));
-      claimable.forEach((crate) => this._applyWithdrawal(balance.claimable, _token, crate));
+      deposited.forEach((crate) => utils.applyDeposit(balance.deposited, _token, crate, currentSeason));
+      withdrawn.forEach((crate) => utils.applyWithdrawal(balance.withdrawn, _token, crate));
+      claimable.forEach((crate) => utils.applyWithdrawal(balance.claimable, _token, crate));
 
       return balance;
     }
@@ -421,7 +247,7 @@ export class Silo {
     /// SETUP
     const whitelist = Silo.sdk.tokens.siloWhitelist;
     const balances = new Map<Token, TokenSiloBalance>();
-    whitelist.forEach((token) => balances.set(token, this._makeTokenSiloBalance()));
+    whitelist.forEach((token) => balances.set(token, utils.makeTokenSiloBalance()));
 
     /// LEDGER
     if (source === DataSource.LEDGER) {
@@ -438,7 +264,7 @@ export class Silo {
       // Attach stalk & seed counts for each crate.
       deposits.forEach((_crates, token) => {
         if (!balances.has(token)) {
-          balances.set(token, this._makeTokenSiloBalance());
+          balances.set(token, utils.makeTokenSiloBalance());
         }
         const state = balances.get(token)!.deposited;
 
@@ -451,30 +277,30 @@ export class Silo {
 
           // Update the total deposited of this token
           // and return a parsed crate object
-          this._applyDeposit(state, token, rawCrate, currentSeason);
+          utils.applyDeposit(state, token, rawCrate, currentSeason);
         }
 
-        this._sortCrates(state);
+        utils.sortCrates(state);
       });
 
       // Handle withdrawals.
       // Split crates into withdrawn and claimable.
       withdrawals.forEach((_crates, token) => {
         if (!balances.has(token)) {
-          balances.set(token, this._makeTokenSiloBalance());
+          balances.set(token, utils.makeTokenSiloBalance());
         }
 
         //
-        const { withdrawn, claimable } = this._parseWithdrawalCrates(token, _crates, seasonBN);
+        const { withdrawn, claimable } = utils.parseWithdrawalCrates(token, _crates, seasonBN);
         const tokenBalance = balances.get(token);
         tokenBalance!.withdrawn = withdrawn;
         tokenBalance!.claimable = claimable;
 
-        this._sortCrates(tokenBalance!.withdrawn);
-        this._sortCrates(tokenBalance!.claimable);
+        utils.sortCrates(tokenBalance!.withdrawn);
+        utils.sortCrates(tokenBalance!.claimable);
       });
 
-      return this._sortTokenMapByWhitelist(balances); // FIXME: sorting is redundant if this is instantiated
+      return utils.sortTokenMapByWhitelist(Silo.sdk.tokens.siloWhitelist, balances); // FIXME: sorting is redundant if this is instantiated
     }
 
     /// SUBGRAPH
@@ -488,7 +314,7 @@ export class Silo {
       const prepareToken = (address: string) => {
         const token = Silo.sdk.tokens.findByAddress(address);
         if (!token) return; // FIXME: unknown token handling
-        if (!balances.has(token)) balances.set(token, this._makeTokenSiloBalance());
+        if (!balances.has(token)) balances.set(token, utils.makeTokenSiloBalance());
         return token;
       };
 
@@ -498,7 +324,7 @@ export class Silo {
         const token = prepareToken(crate.token);
         if (!token) return;
         const state = balances.get(token)!.deposited;
-        this._applyDeposit(state, token, crate, currentSeason);
+        utils.applyDeposit(state, token, crate, currentSeason);
       };
 
       // Handle withdrawals.
@@ -508,14 +334,14 @@ export class Silo {
         const token = prepareToken(crate.token);
         if (!token) return;
         const state = balances.get(token)![key];
-        this._applyWithdrawal(state, token, crate);
+        utils.applyWithdrawal(state, token, crate);
       };
 
       deposited.forEach(handleDeposit);
       withdrawn.forEach(handleWithdrawal("withdrawn"));
       claimable.forEach(handleWithdrawal("claimable"));
 
-      return this._sortTokenMapByWhitelist(balances);
+      return utils.sortTokenMapByWhitelist(Silo.sdk.tokens.siloWhitelist, balances);
     }
 
     throw new Error(`Unsupported source: ${source}`);
@@ -599,72 +425,6 @@ export class Silo {
     return Silo.sdk.contracts.beanstalk.balanceOfGrownStalk(account).then((v) => Silo.sdk.tokens.STALK.fromBlockchain(v));
   }
 
-  //////////////////////// Crates ////////////////////////
-
-  pickCrates(
-    crates: Crate<TokenValue>[],
-    token: Token,
-    amount: BigNumber | TokenValue,
-    sort: CrateSortFn = (crates) => sortCratesBySeason(crates, "desc")
-  ) {
-    const sortedCrates = sort(crates);
-    const seasons: string[] = [];
-    const amounts: string[] = [];
-    let remaining = amount instanceof TokenValue ? TokenValue.from(amount) : TokenValue.fromBlockchain(amount, token.decimals);
-    sortedCrates.some((crate) => {
-      const thisAmount = crate.amount.gt(remaining) ? crate.amount.sub(remaining) : crate.amount;
-      seasons.push(crate.season.toString());
-      amounts.push(thisAmount.toBlockchain());
-      remaining = remaining.sub(thisAmount);
-      return remaining.eq(0); // done
-    });
-    if (!remaining.eq(0)) throw new Error("Not enough amount in crates");
-    return { seasons, amounts };
-  }
-
-  sumDeposits(token: ERC20Token, crates: DepositCrate[]) {
-    return crates.reduce(
-      (prev, curr) => {
-        prev.amount = prev.amount.add(curr.amount);
-        prev.stalk = prev.stalk.add(curr.stalk);
-        prev.seeds = prev.seeds.add(curr.seeds);
-        prev.bdv = prev.bdv.add(curr.bdv);
-        return prev;
-      },
-      {
-        amount: token.amount(0),
-        stalk: Silo.sdk.tokens.STALK.amount(0),
-        seeds: Silo.sdk.tokens.SEEDS.amount(0),
-        bdv: Silo.sdk.tokens.BEAN.amount(0)
-      }
-    );
-  }
-
-  async bdv(_token: Token, _amount?: TokenValue) {
-    return Silo.sdk.contracts.beanstalk
-      .bdv(_token.address, (_amount || _token.amount(1)).toBlockchain())
-      .then((v) => Silo.sdk.tokens.BEAN.fromBlockchain(v));
-  }
-
-  //////////////////////// ACTION: Claim Rewards ////////////////////////
-  /**
-   * Mowing adds Grown Stalk to stalk balance
-   * @param _account
-   */
-  async mow(_account?: string): Promise<ContractTransaction> {
-    const account = _account ? _account : await Silo.sdk.getAccount();
-    return Silo.sdk.contracts.beanstalk.update(account);
-  }
-
-  /**
-   * Claims Earned Beans, Earned Stalk, Plantable Seeds and also mows any Grown Stalk
-   */
-  async plant(): Promise<ContractTransaction> {
-    return Silo.sdk.contracts.beanstalk.plant();
-  }
-
-  //////////////////////// Permits ////////////////////////
-
   /**
    * Created typed permit data to authorize `spender` to transfer
    * the `owner`'s deposit balance of `token`.
@@ -691,11 +451,11 @@ export class Silo {
   ): Promise<EIP712TypedData<DepositTokenPermitMessage>> {
     const deadline = _deadline || MAX_UINT256;
     const [domain, nonce] = await Promise.all([
-      this._getEIP712Domain(),
+      permitUtils.getEIP712Domain(),
       _nonce || Silo.sdk.contracts.beanstalk.depositPermitNonces(owner).then((nonce) => nonce.toString())
     ]);
 
-    return this._createTypedDepositTokenPermitData(domain, {
+    return permitUtils.createTypedDepositTokenPermitData(domain, {
       owner,
       spender,
       token,
@@ -735,11 +495,11 @@ export class Silo {
 
     const deadline = _deadline || MAX_UINT256;
     const [domain, nonce] = await Promise.all([
-      this._getEIP712Domain(),
+      permitUtils.getEIP712Domain(),
       _nonce || Silo.sdk.contracts.beanstalk.depositPermitNonces(owner).then((nonce) => nonce.toString())
     ]);
 
-    return this._createTypedDepositTokensPermitData(domain, {
+    return permitUtils.createTypedDepositTokensPermitData(domain, {
       owner,
       spender,
       tokens,
@@ -748,53 +508,4 @@ export class Silo {
       deadline
     });
   }
-
-  /**
-   * Get the EIP-712 domain for the Silo.
-   * @note applies to both `depositToken` and `depositTokens` permits.
-   */
-  private async _getEIP712Domain() {
-    return {
-      name: "SiloDeposit",
-      version: "1",
-      // FIXME: switch to below after protocol patch
-      // chainId: (await Silo.sdk.provider.getNetwork()).chainId,
-      chainId: 1,
-      verifyingContract: "0xc1e088fc1323b20bcbee9bd1b9fc9546db5624c5"
-    };
-  }
-
-  private _createTypedDepositTokenPermitData = (domain: EIP712Domain, message: DepositTokenPermitMessage) => ({
-    types: {
-      EIP712Domain: Permit.EIP712_DOMAIN,
-      Permit: [
-        { name: "owner", type: "address" },
-        { name: "spender", type: "address" },
-        { name: "token", type: "address" },
-        { name: "value", type: "uint256" },
-        { name: "nonce", type: "uint256" },
-        { name: "deadline", type: "uint256" }
-      ]
-    },
-    primaryType: "Permit",
-    domain,
-    message
-  });
-
-  private _createTypedDepositTokensPermitData = (domain: EIP712Domain, message: DepositTokensPermitMessage) => ({
-    types: {
-      EIP712Domain: Permit.EIP712_DOMAIN,
-      Permit: [
-        { name: "owner", type: "address" },
-        { name: "spender", type: "address" },
-        { name: "tokens", type: "address[]" },
-        { name: "values", type: "uint256[]" },
-        { name: "nonce", type: "uint256" },
-        { name: "deadline", type: "uint256" }
-      ]
-    },
-    primaryType: "Permit",
-    domain,
-    message
-  });
 }
